+++ frontend/src/components/StudentDashboard.js
@@ -23,6 +23,7 @@ import WarningIcon from "@mui/icons-material/Warning";
 import ScheduleIcon from "@mui/icons-material/Schedule";
 import PeopleIcon from "@mui/icons-material/People";
 import { UserInfoContext } from "../stores/UserInfoStore";
+import SelectInterviewDate from "./Dashboard/SelectInterviewDate";
 
 const StudentDashBoard = () => {
   const listState = [
@@ -36,8 +37,6 @@ const StudentDashBoard = () => {
     "STAGE TROUVÉE",
     "EN ATTENTE DE RÉPONSE",
   ];
-
-  const currentDate = new Date();
 
   const [profile, setProfile] = useState({
     id: "",
@@ -56,9 +55,7 @@ const StudentDashBoard = () => {
 
   const [isDisabled, setIsDisabled] = useState(true);
   const [isStatusUpdated, setIsStatusUpdated] = useState(false);
-  const [isDateValid, setIsDateValid] = useState(true);
   const [userInfo] = useContext(UserInfoContext);
-  const [isInterviewDateUpdated, setIsInterviewDateUpdated] = useState(false);
 
   useEffect(() => {
     const getProfile = () => {
@@ -88,7 +85,6 @@ const StudentDashBoard = () => {
   const showAndHideStatusUpdateSuccessMsg = () => {
     setTimeout(() => {
       setIsStatusUpdated(false);
-      setIsDateValid(true);
     }, 3000);
   };
 
@@ -109,39 +105,6 @@ const StudentDashBoard = () => {
       });
   };
 
-  const showAndHideInterviewDateErrorMsg = () => {
-    setIsDateValid(false);
-    setTimeout(() => {
-      setIsDateValid(true);
-    }, 3000);
-  };
-
-  const showAndHideInterviewUpdateSuccessMsg = () => {
-    setIsInterviewDateUpdated(true);
-
-    setTimeout(() => {
-      setIsInterviewDateUpdated(false);
-    }, 3000);
-  };
-
-  const updateInterviewDate = (date) => {
-    axios({
-      method: "PATCH",
-      url: "http://localhost:8080/student/updateInterviewDate/" + date,
-      headers: {
-        Authorization: userInfo.jwt,
-      },
-      responseType: "json",
-    })
-      .then(() => {
-        showAndHideInterviewUpdateSuccessMsg();
-      })
-      .catch((error) => {
-        showAndHideInterviewDateErrorMsg();
-        console.error(error);
-      });
-  };
-
   const hasInternship = () => {
     return profile.studentState === listState[1];
   };
@@ -150,25 +113,6 @@ const StudentDashBoard = () => {
     setProfile({ ...profile, studentState: $event.target.value });
     updateStudentStatus();
     setIsDisabled(true);
-  };
-
-  const handleChangeDate = ($event) => {
-    const value = $event.target.value;
-    const dateValues = value.split("-");
-
-    if (
-      (currentDate.getFullYear() === parseInt(dateValues[0]) &&
-        parseInt(dateValues[1]) > currentDate.getMonth() + 1) ||
-      (parseInt(dateValues[2], 10) >= currentDate.getDate() &&
-        parseInt(dateValues[1]) === currentDate.getMonth() + 1 &&
-        !hasInternship())
-    ) {
-      setProfile({ ...profile, closestInterviewDate: value });
-      updateInterviewDate(value);
-      setIsDateValid(true);
-    } else {
-      showAndHideInterviewDateErrorMsg();
-    }
   };
 
   const fadeIn = {
@@ -267,77 +211,13 @@ const StudentDashBoard = () => {
                       <TodayIcon />
                     </Typography>
 
-                    {!hasInternship() && profile.hasValidCv ? (
-                      <TextField
-                        sx={{
-                          border: "1px white",
-                          boxShadow: 4,
-                          borderRadius: 2,
-                          display: "flex",
-                          m: 2,
-                        }}
-                        InputLabelProps={{
-                          shrink: true,
-                        }}
-                        label="Date d'entrevue la plus proche"
-                        value={profile.closestInterviewDate}
-                        onChange={handleChangeDate}
-                        type={"date"}
-                      />
-                    ) : (
-                      <motion.div
-                        variants={fadeIn}
-                        initial="hidden"
-                        animate="show"
-                      >
-                        <Typography
-                          variant="subtitle1"
-                          sx={{
-                            color: "gray",
-                            textAlign: "center",
-                            alignItems: "center",
-                            m: 2,
-                          }}
-                        >
-                          {
-                            "L'ajout d'entrevue n'est pas disponible pour l'instant"
-                          }
-                          <ScheduleIcon />
-                        </Typography>
-                      </motion.div>
-                    )}
-
-                    {isInterviewDateUpdated ? (
-                      <motion.div
-                        variants={fadeIn}
-                        initial="hidden"
-                        animate="show"
-                      >
-                        <Typography
-                          variant="subtitle1"
-                          sx={{ color: "green", textAlign: "center" }}
-                        >
-                          {"DATE D'ENTREVUE AJOUTÉE"}
-                          <PublishedWithChangesIcon />
-                        </Typography>
-                      </motion.div>
-                    ) : null}
-
-                    {!isDateValid ? (
-                      <motion.div
-                        variants={fadeIn}
-                        initial="hidden"
-                        animate="show"
-                      >
-                        <Typography
-                          variant="subtitle1"
-                          sx={{ color: "red", textAlign: "center" }}
-                        >
-                          {"DATE D'ENTREVUE INVALIDE"}
-                          <WarningIcon />
-                        </Typography>
-                      </motion.div>
-                    ) : null}
+                    <SelectInterviewDate
+                      hasInternship={hasInternship}
+                      profile={profile}
+                      setProfile={setProfile}
+                      jwt={userInfo.jwt}
+                      fadeIn={fadeIn}
+                    />
                   </Grid>
                 </Grid>
                 <Grid
